pyproject.toml
[build-system]
requires = [
    "setuptools>=63.0",
    "trove-classifiers",
    "wheel",
]

build-backend = "setuptools.build_meta"

[project]
name = "flickr-tools"
version = "0.0.1"
description = "Download/process Flickr content"
authors = [
    { name = "Skip Montanaro", email = "skip.montanaro@gmail.com" },
]
requires-python = ">=3.11"

dependencies = [
    "python-dateutil>=2.8.2",
    "flickr-api>=0.7.6",
    "regex>=2022.10.31",
    "requests>=2.28.1",
]

[tool.setuptools.packages.find]
include = ["flickr_tools"]

[project.scripts]
    flickr-download = "flickr_tools.flickr_download:main"
